requirements.txt
streamlit
paho-mqtt<2.0.0